refactor(imports): reorder and simplify all import statements
This commit applies a consistent ordering to all `import` statements throughout the `src/` directory to improve code readability and maintainability.

The new order is:
1. Local imports (from the same directory).
2. Parent and sibling directory imports, grouped logically by the main project folders (core, features, shared).

Additionally, any unnecessarily complex paths (e.g., using `../` when `./` would suffice) have been simplified. No functional changes are included.
