+++ src/core/documents/actor.js
@@ -1,22 +1,22 @@
 /* global Actor, Application, CONFIG, CONST, Dialog, Die, foundry, fromUuid, fromUuidSync, game, Hooks, Roll, TextEditor, Token, ui */
 
-import { COC7 } from '../config.js'
 import CoC7ActiveEffect from './active-effect.js'
-import { CoC7Check } from '../../core/check.js'
+import { CoC7Item } from './item.js'
+import { CoC7Check } from '../check.js'
+import { CoC7ChatMessage } from '../coc7-chat-message.js'
+import { COC7 } from '../config.js'
+import { CoC7ConCheck } from '../../features/check-resolution/concheck.js'
 import { CoC7MeleeInitiator } from '../../features/combat/melee-initiator.js'
 import { CoC7RangeInitiator } from '../../features/combat/range-combat.js'
-import { CoC7ConCheck } from '../../features/check-resolution/concheck.js'
-import { CoC7ChatMessage } from '../coc7-chat-message.js'
-import { RollDialog } from '../../shared/ui-dialogs/roll-dialog.js'
 import { ExperiencePackageDialog } from '../../features/item-experience-package/apps/experience-package-dialog.js'
-import { chatHelper } from '../../shared/dice/helper.js'
 import { CoC7Dice } from '../../shared/dice/dice.js'
-import { CoC7Item } from '../documents/item.js'
+import { chatHelper } from '../../shared/dice/helper.js'
 import { AverageRoll } from '../../shared/ui-dialogs/average-roll.js'
-import { SkillSelectDialog } from '../../shared/ui-dialogs/skill-selection-dialog.js'
-import { PointSelectDialog } from '../../shared/ui-dialogs/point-selection-dialog.js'
-import { CharacSelectDialog } from '../../shared/ui-dialogs/char-selection-dialog.js'
 import { CharacRollDialog } from '../../shared/ui-dialogs/char-roll-dialog.js'
+import { CharacSelectDialog } from '../../shared/ui-dialogs/char-selection-dialog.js'
+import { PointSelectDialog } from '../../shared/ui-dialogs/point-selection-dialog.js'
+import { RollDialog } from '../../shared/ui-dialogs/roll-dialog.js'
+import { SkillSelectDialog } from '../../shared/ui-dialogs/skill-selection-dialog.js'
 import { SkillSpecSelectDialog } from '../../shared/ui-dialogs/skill-spec-select-dialog.js'
 import { SkillSpecializationSelectDialog } from '../../shared/ui-dialogs/skill-specialization-select-dialog.js'
 import { SkillValueDialog } from '../../shared/ui-dialogs/skill-value-dialog.js'
