+++ src/core/index.js
@@ -3,9 +3,6 @@ import '../../styles/system/index.less'
 import './polyfill.js'
 /* // FoundryVTT V10 */
 // import * as DiceBot from './dicebot.js'
-import { initECC } from '../shared/chatcardlib/src/chatcardlib.js'
-import { ChaseObstacleCard } from '../features/chase/chat/chase-obstacle.js'
-
 import ChangeSidebarTab from './hooks/change-sidebar-tab.js'
 import ChatMessage from './hooks/chat-message.js'
 import CloseActorSheet from './hooks/close-actor-sheet.js'
@@ -16,25 +13,16 @@ import DiceSoNiceReady from './hooks/dice-so-nice-ready.js'
 import DrawNote from './hooks/draw-note.js'
 import DropActorSheetData from './hooks/drop-actor-sheet-data.js'
 import DropCanvasData from './hooks/drop-canvas-data.js'
-import GetHeaderControlsJournalEntrySheet from '../features/coc-id-system/hooks/get-header-controls-journal-entry-sheet.js'
-import GetHeaderControlsMacroConfig from '../features/coc-id-system/hooks/get-header-controls-macro-config.js'
-import GetHeaderControlsPlaylistConfig from '../features/coc-id-system/hooks/get-header-controls-playlist-config.js'
-import GetHeaderControlsRollTableSheet from '../features/coc-id-system/hooks/get-header-controls-roll-table-sheet.js'
-import GetHeaderControlsSceneConfig from '../features/coc-id-system/hooks/get-header-controls-scene-config.js'
 import GetSceneControlButtons from './hooks/get-scene-control-buttons.js'
 import HotbarDrop from './hooks/hotbar-drop.js'
 import Init from './hooks/init.js'
 import PopOutLoaded from './hooks/popout-loaded.js'
 import Ready from './hooks/ready.js'
-import RenderActorDirectory from '../features/investigator-wizard/hooks/render-actor-directory.js'
-import RenderActorSheet from '../features/actor-character/hooks/render-actor-sheet.js'
 import RenderChatLog from './hooks/render-chat-log.js'
 import RenderChatMessage from './hooks/render-chat-message.js'
 import RenderCoC7DirectoryPicker from './hooks/render-coc7-directory-picker.js'
-import RenderCombatTracker from '../features/combat/hooks/render-combat-tracker.js'
 import RenderDialog from './hooks/render-dialog.js'
 import RenderGamePause from './hooks/render-game-pause.js'
-import RenderItemSheet from '../features/actor-character/hooks/render-item-sheet.js'
 import RenderJournalEntryPageTextSheet from './hooks/render-journal-entry-page-text-sheet.js'
 import RenderJournalEntrySheet from './hooks/render-journal-entry-sheet.js'
 import RenderJournalSheet from './hooks/render-journal-sheet.js'
@@ -49,11 +37,21 @@ import RenderSettingsConfig from './hooks/render-settings-config.js'
 import Setup from './hooks/setup.js'
 import SocketlibReady from './hooks/socketlib-ready.js'
 import UpdateChatMessage from './hooks/update-chat-message.js'
-
-import CloseCoC7ChaseSheet from '../features/chase/hooks/close-coc7-chase-sheet.js'
-import RenderCoC7ChaseSheet from '../features/chase/hooks/render-coc7-chase-sheet.js'
+import RenderActorSheet from '../features/actor-character/hooks/render-actor-sheet.js'
+import RenderItemSheet from '../features/actor-character/hooks/render-item-sheet.js'
 import RenderCoC7CreatureSheet from '../features/actor-creature/hooks/render-coc7-creature-sheet.js'
 import RenderCoC7NPCSheet from '../features/actor-npc/hooks/render-coc7-npc-sheet.js'
+import { ChaseObstacleCard } from '../features/chase/chat/chase-obstacle.js'
+import CloseCoC7ChaseSheet from '../features/chase/hooks/close-coc7-chase-sheet.js'
+import RenderCoC7ChaseSheet from '../features/chase/hooks/render-coc7-chase-sheet.js'
+import GetHeaderControlsJournalEntrySheet from '../features/coc-id-system/hooks/get-header-controls-journal-entry-sheet.js'
+import GetHeaderControlsMacroConfig from '../features/coc-id-system/hooks/get-header-controls-macro-config.js'
+import GetHeaderControlsPlaylistConfig from '../features/coc-id-system/hooks/get-header-controls-playlist-config.js'
+import GetHeaderControlsRollTableSheet from '../features/coc-id-system/hooks/get-header-controls-roll-table-sheet.js'
+import GetHeaderControlsSceneConfig from '../features/coc-id-system/hooks/get-header-controls-scene-config.js'
+import RenderCombatTracker from '../features/combat/hooks/render-combat-tracker.js'
+import RenderActorDirectory from '../features/investigator-wizard/hooks/render-actor-directory.js'
+import { initECC } from '../shared/chatcardlib/src/chatcardlib.js'
 
 Hooks.once('init', Init)
 Hooks.once('ready', Ready)
