+++ src/core/sheets/actor-sheet-base.js
@@ -1,20 +1,20 @@
 /* global $, ChatMessage, CONST, Dialog, FormData, foundry, game, Hooks, Roll, TextEditor, ui */
-import { addCoCIDSheetHeaderButton } from '../../features/coc-id-system/coc-id-button.js'
-import { RollDialog } from '../../shared/ui-dialogs/roll-dialog.js'
-import { CoC7ChatMessage } from '../coc7-chat-message.js'
 import { CoC7Check } from '../../core/check.js'
-import { CoC7ContentLinkDialog } from '../../features/link-creation/coc7-content-link-dialog.js'
+import { CoC7ChatMessage } from '../coc7-chat-message.js'
 import { COC7 } from '../config.js'
+import CoC7ActiveEffect from '../../core/documents/active-effect.js'
 import { CoCActor } from '../documents/actor.js'
 import { CoC7Item } from '../documents/item.js'
+import { CoC7ConCheck } from '../../features/check-resolution/concheck.js'
+import { DamageCard } from '../../features/combat/chat/damage.js'
 import { CoC7MeleeInitiator } from '../../features/combat/melee-initiator.js'
 import { CoC7RangeInitiator } from '../../features/combat/range-combat.js'
-import { CoC7ConCheck } from '../../features/check-resolution/concheck.js'
-import { chatHelper, isCtrlKey } from '../../shared/dice/helper.js'
+import { addCoCIDSheetHeaderButton } from '../../features/coc-id-system/coc-id-button.js'
+import { CoC7ContentLinkDialog } from '../../features/link-creation/coc7-content-link-dialog.js'
 import { CoC7Link } from '../../features/link-creation/coc7-link.js'
-import { DamageCard } from '../../features/combat/chat/damage.js'
-import CoC7ActiveEffect from '../../core/documents/active-effect.js'
+import { chatHelper, isCtrlKey } from '../../shared/dice/helper.js'
 import { CoC7ContextMenu } from '../../shared/ui-components/context-menu.js'
+import { RollDialog } from '../../shared/ui-dialogs/roll-dialog.js'
 import { CoC7Utilities } from '../../shared/utilities.js'
 
 /**
