+++ src/core/system/register-sheets.js
@@ -1,24 +1,24 @@
 /* global Actors, foundry, Items */
+import { CoC7CharacterSheet } from '../../features/actor-character/sheet.js'
+import { CoC7CharacterSheetMinimized } from '../../features/actor-character/sheet-minimized.js'
+import { CoC7CharacterSheetV3 } from '../../features/actor-character/sheet-v3.js'
+import { CoC7ContainerSheet } from '../../features/actor-container/sheet.js'
+import { CoC7CreatureSheet } from '../../features/actor-creature/sheet.js'
+import { CoC7NPCSheet } from '../../features/actor-npc/sheet.js'
+import { CoC7VehicleSheet } from '../../features/actor-vehicle/sheet.js'
+import { CoC7ChaseSheet } from '../../features/chase/sheet.js'
 import { CoC7ArchetypeSheet } from '../../features/item-archetype/sheet.js'
 import { CoC7ArmorSheet } from '../../features/item-armor/sheet.js'
 import { CoC7BookSheet } from '../../features/item-book/sheet.js'
-import { CoC7CharacterSheet } from '../../features/actor-character/sheet.js'
-import { CoC7CharacterSheetV3 } from '../../features/actor-character/sheet-v3.js'
-import { CoC7CharacterSheetMinimized } from '../../features/actor-character/sheet-minimized.js'
-import { CoC7ContainerSheet } from '../../features/actor-container/sheet.js'
-import { CoC7ChaseSheet } from '../../features/chase/sheet.js'
-import { CoC7CreatureSheet } from '../../features/actor-creature/sheet.js'
 import { CoC7ExperiencePackageSheet } from '../../features/item-experience-package/sheet.js'
 import { CoC7ItemSheet } from '../../features/item-generic/sheet.js'
 import { CoC7ItemSheetV2 } from '../../features/item-generic/sheet-v2.js'
-import { CoC7NPCSheet } from '../../features/actor-npc/sheet.js'
 import { CoC7OccupationSheet } from '../../features/item-occupation/sheet.js'
 import { CoC7SetupSheet } from '../../features/item-setup/sheet.js'
 import { CoC7SkillSheet } from '../../features/item-skill/sheet.js'
 import { CoC7SpellSheet } from '../../features/item-spell/sheet.js'
 import { CoC7StatusSheet } from '../../features/item-status/sheet.js'
 import { CoC7TalentSheet } from '../../features/item-talent/sheet.js'
-import { CoC7VehicleSheet } from '../../features/actor-vehicle/sheet.js'
 import { CoC7WeaponSheet } from '../../features/item-weapon/sheet.js'
 
 export function registerSheets () {
