+++ src/features/investigator-wizard/hooks/render-actor-directory.js
@@ -1,6 +1,6 @@
 /* global CONST, foundry, game */
-import { CoC7ActorImporterDialog } from '../../actor-importer/actor-importer-dialog.js'
 import { CoC7InvestigatorWizard } from '../investigator-wizard.js'
+import { CoC7ActorImporterDialog } from '../../actor-importer/actor-importer-dialog.js'
 
 export default function (application, html, data) {
   // Allow Investigator Wizard
