+++ src/features/item-status/sheet.js
@@ -1,7 +1,7 @@
 /* global foundry, game, TextEditor */
 import { addCoCIDSheetHeaderButton } from '../coc-id-system/coc-id-button.js'
-import CoC7ActiveEffect from '../../core/documents/active-effect.js'
 import { COC7 } from '../../core/config.js'
+import CoC7ActiveEffect from '../../core/documents/active-effect.js'
 
 /**
  * Extend the basic ItemSheet with some very simple modifications
